// angular-rms/src/app/shared/model/employee.model.ts
export class Employee {
    Id: number;
    firstName: string;
    lastName: string;
    gender: string;
    dob: string;
    nationality: string;
    maritalStatus: string;
    phone: string;
    subDivision: string;
    status: string;
    suspendDate: string;
    hiredDate: string;
    grade: string;
    division: string;
    email: string;
    location: Location;
    photo: string;

    Employee() {}
}